Fixed error resulting from np.sqrt of large ints (replaced by floats) in keyVal

+++ RNAseqAnalysis.py
@@ -1006,7 +1006,8 @@ def keyVal(key, d, meth):
     switch = {'product': lambda TP, FN, TN, FP: TP/(TP+FN) * TN/(TN+FP),
               'Youden': lambda TP, FN, TN, FP: TP/(TP+FN) + TN/(TN+FP) - 1,
               'MCC': lambda TP, FN, TN, FP:
-                  (TP*TN - FP*FN) / np.sqrt((TP+FN)*(TP+FP)*(TN+FP)*(TN+FN)) }
+                  (TP*TN - FP*FN) / 
+                  np.sqrt(int((TP + FN) * (TP + FP) * (TN + FP) * (TN + FN))) }
     val = switch[meth](*d[key][:4])
     if np.isnan(val):
         val = 0
